refactor(build): Update linker configuration and dependency management
- Modify .cargo/config.toml to use clang and lld linker
- Remove mold linker configuration for x86_64-unknown-linux-gnu
- Comment out macOS-specific linker configuration
- Update Cargo.lock with new dependency versions
- Add several new dependencies to improve project configuration and functionality
- Prepare project for more robust build and linking performance
