+++ Cargo.toml
@@ -24,3 +24,11 @@ path = "src/lib.rs"
 [dev-dependencies]
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
 tokio-test = "0.4"
+
+[dependencies.sqlx]
+version = "0.8.6"
+default-features = false
+features = ["runtime-tokio-rustls", "macros",
+"postgres", "uuid", "chrono", "migrate"
+]
+
