+++ src/database/mod.rs
@@ -1,1 +1,3 @@
 pub mod data_base;
+
+pub use data_base::{FormUsers, UserDB};
