+++ src/handlers/health.rs
@@ -2,6 +2,7 @@ use axum::{response::IntoResponse, http::StatusCode};
 
 
 ///# Health check message
-pub async fn health_check() -> impl IntoResponse {
+pub async fn health_check() -> impl IntoResponse 
+{
     StatusCode::OK
 }
